feat(de forms): implement validation with zod
