+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-10-25T11:43:32.223Z\n"
-"PO-Revision-Date: 2023-10-25T11:43:32.223Z\n"
+"POT-Creation-Date: 2023-11-16T09:05:06.098Z\n"
+"PO-Revision-Date: 2023-11-16T09:05:06.098Z\n"
 
 msgid "schemas"
 msgstr "schemas"
@@ -78,14 +78,14 @@ msgstr "Aggregation level(s)"
 msgid "Category combo"
 msgstr "Category combo"
 
-msgid "Default (none)"
-msgstr "Default (none)"
+msgid "None"
+msgstr "None"
 
 msgid "Filter legend sets"
 msgstr "Filter legend sets"
 
-msgid "None"
-msgstr "None"
+msgid "<No value>"
+msgstr "<No value>"
 
 msgid "Option set"
 msgstr "Option set"
@@ -685,6 +685,21 @@ msgstr ""
 msgid "Aggregation levels"
 msgstr "Aggregation levels"
 
+msgid ""
+"By default, the aggregation will start at the lowest assigned organisation "
+"unit. If you for example select \"Chiefdom\", it means that \"Chiefdom\", "
+"\"District\" and \"National\" aggregates use \"Chiefdom\" (the highest "
+"aggregation level available) as the data source, and PHU data will not be "
+"included. PHU will still be available for the PHU level, but not included "
+"in the aggregations to the levels above."
+msgstr ""
+"By default, the aggregation will start at the lowest assigned organisation "
+"unit. If you for example select \"Chiefdom\", it means that \"Chiefdom\", "
+"\"District\" and \"National\" aggregates use \"Chiefdom\" (the highest "
+"aggregation level available) as the data source, and PHU data will not be "
+"included. PHU will still be available for the PHU level, but not included "
+"in the aggregations to the levels above."
+
 msgid "Custom fields for your DHIS2 instance"
 msgstr "Custom fields for your DHIS2 instance"
 
@@ -730,6 +745,12 @@ msgstr "Option set comment"
 
 msgid "Choose a set of predefined comment for data entry"
 msgstr "Choose a set of predefined comment for data entry"
+
+msgid "Required"
+msgstr "Required"
+
+msgid "The value is to long. You can use up to 255 characters"
+msgstr "The value is to long. You can use up to 255 characters"
 
 msgid "Metadata management"
 msgstr "Metadata management"
