+++ src/pages/dataElements/Edit.tsx
@@ -15,7 +15,11 @@ import { JsonPatchOperation } from '../../types'
 import { Attribute, DataElement } from '../../types/generated'
 import { createJsonPatchOperations } from './edit/'
 import classes from './Edit.module.css'
-import { DataElementFormFields, useCustomAttributesQuery } from './form'
+import {
+    DataElementFormFields,
+    useCustomAttributesQuery,
+    validate,
+} from './form'
 import type { FormValues } from './form'
 
 type FinalFormFormApi = FormApi<FormValues>
@@ -170,7 +174,11 @@ export const Component = () => {
                 queryResponse={customAttributesQuery}
                 label={i18n.t('Custom attributes')}
             >
-                <Form onSubmit={onSubmit} initialValues={initialValues}>
+                <Form
+                    onSubmit={onSubmit}
+                    validate={validate}
+                    initialValues={initialValues}
+                >
                     {({ handleSubmit, submitting, submitError }) => (
                         <form onSubmit={handleSubmit}>
                             <FormContents
